+++ docs/history.rst
@@ -4,7 +4,14 @@
 History
 =======
 
+Version 0.2.0 - Unreleased
+--------------------------
+
+* Added class ``MakevokePrintOut`` for printout methods with tests;
+
+
 Version 0.1.0 - Unreleased
 --------------------------
 
-* First commit.
+Initial commit with very basic stuff.
+